Update the default landing page promo (behind switch)
