+++ test/services/WeeklyPickerTest.scala
@@ -1,4 +1,4 @@
-package views.support
+package services
 
 import com.gu.i18n.{Country, CountryGroup}
 import org.specs2.mock.Mockito
